pyproject.toml
[project]
name = "aiglasses-navigation"
version = "0.1.0"
description = "AI智能盲人眼镜导航系统 - 视障人士智能导航辅助"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
    "python-multipart==0.0.6",
    "opencv-python==4.8.1.78",
    "opencv-contrib-python==4.8.1.78",
    "numpy>=1.24.3,<2",
    "pillow==10.1.0",
    "pyaudio==0.2.14",
    "pydub==0.25.1",
    "pygame==2.5.2",
    "dashscope==1.14.1",
    "python-dotenv>=1.0.0",
    "ncnn>=1.0.20260114",
    "pnnx>=20260409",
]

[dependency-groups]
dev = [
    "platformio>=6.1.19",
]
